backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
